Refactoring to avoid creating the "replaceP"-function multiple times (createrMarkerClass-function with Promise-constructor as argument) Include extract "lib/*"-files into Extract utility functions to separate file.
