+++ lib/replaceP.js
@@ -5,6 +5,8 @@
  * Released under the MIT license.
  */
 
+'use strict'
+
 /**
  * Creates a String.prototype.replace function with support for Promises using a specific Promise constructor.
  *
@@ -12,7 +14,7 @@
  * @returns {function(string, RegExp, function)} an equivalent for String.prototype.replace which
  *  can handle promises
  */
-module.exports = function (Promise) {
+module.exports = function createReplaceP(Promise) {
   var deepAPlus = require('deep-aplus')(Promise)
   /**
    * Similar to String.prototype.replace, but it returns a promise instead of a string
